hgnn_project_copy/requirements.txt
# Compatible Torch + CUDA version
torch==2.5.1+cu121
torchvision==0.16.1+cu121
torchaudio==2.5.1
# pip install --index-url https://download.pytorch.org/whl/cu121

# PyG & dependencies
torch-scatter==2.1.2
torch-sparse==0.6.18
torch-cluster==1.6.1
torch-geometric==2.5.3

# Graph + point cloud utils
networkx
trimesh
pyglet
tqdm

# Scientific computing
numpy
scipy
scikit-learn
matplotlib

# Config and serialization
pyyaml
